Scripts to migrate VMs from external iSCSI to ONE
Initial import.

This is a set of scripts to support the process of migration of virtual
machines to OpenNebula.

The scripts create a VM in OpenNabula and persistent images for virtual
disk(s), copy the disk images from a source iSCSI SAN to the created
images on StorPool and start the VM.